requirements.txt
scrapy
requests
progressbar2